split up the classes (#2824)
