+++ app/domain/people/duplicate_locator.rb
@@ -20,27 +20,27 @@ class People::DuplicateLocator
 
   def run
     @scope.find_each do |person|
-      duplicate = find_duplicate(person)
+      duplicate_id = find_duplicate_id(person)
 
-      next unless duplicate
+      next unless duplicate_id
 
       # Sorting by id to only allow a single PersonDuplicate entry per Person combination
-      person_1, person_2 = [person, duplicate].sort_by(&:id)
+      person_1, person_2 = [person.id, duplicate_id].sort
 
-      PersonDuplicate.find_or_create_by!(person_1: person_1, person_2: person_2)
+      PersonDuplicate.find_or_create_by!(person_1_id: person_1, person_2_id: person_2)
     end
   end
 
   private
 
-  def find_duplicate(person)
+  def find_duplicate_id(person)
     criterion = DUPLICATION_ATTRS.index_with { |attr| person[attr] }
-    duplicate = person_duplicate_finder.find(criterion)
+    duplicate_ids = person_duplicate_detector.find_people_ids(criterion)
 
-    duplicate unless person == duplicate
+    duplicate_ids.first unless person.id == duplicate_ids.first
   end
 
-  def person_duplicate_finder
-    @person_duplicate_finder ||= Import::PersonDuplicateFinder.new
+  def person_duplicate_detector
+    @person_duplicate_detector ||= Import::PersonDuplicateDetector.new
   end
 end
